Move XRM database to x11rb-protocol
This commit moves the code in x11rb::resource_manager to x11rb-protocol.
The only thing requiring I/O in this module is loading the resource
management database. The necessary X11 requests are abstracted out so
that any caller can do so. The file I/O should be quick and is left
as-is.

A new x11rb::resource_manager module is added that provides convenience
functions to load the resource management database from a given &impl
Connection. These functions "close the gap" that is created by the move
to x11rb-protocol and provide exactly the same behaviour as we had
before.
