+++ x11rb-protocol/Cargo.toml
@@ -19,6 +19,10 @@ keywords = ["xcb", "X11"]
 nix = "0.23"
 
 [features]
+# Enable utility functions in `x11rb::resource_manager` for querying the
+# resource databases.
+resource_manager = []
+
 # Features to enable individual X11 extensions
 composite = ["xfixes"]
 damage = ["xfixes"]
