+++ x11rb/examples/simple_window.rs
@@ -4,7 +4,7 @@ use x11rb::connection::Connection;
 use x11rb::cursor::Handle as CursorHandle;
 use x11rb::protocol::xproto::*;
 use x11rb::protocol::Event;
-use x11rb::resource_manager::Database;
+use x11rb::resource_manager::new_from_default;
 use x11rb::wrapper::ConnectionExt as _;
 
 fn main() {
@@ -17,7 +17,7 @@ fn main() {
     let screen = &conn.setup().roots[screen_num];
     let win_id = conn.generate_id().unwrap();
     let gc_id = conn.generate_id().unwrap();
-    let resource_db = Database::new_from_default(conn).unwrap();
+    let resource_db = new_from_default(conn).unwrap();
     let cursor_handle = CursorHandle::new(conn, screen_num, &resource_db).unwrap();
 
     let wm_protocols = conn.intern_atom(false, b"WM_PROTOCOLS").unwrap();
